final touch

+++ app/src/main/java/com/example/beamagayver/data/PrefManager.java
@@ -54,7 +54,7 @@ public class PrefManager {
         return prefs.getBoolean(key, false);
     }
 
-    public long readLong(String key) {
-        return prefs.getLong(key, 0);
+    public int readInt(String key) {
+        return prefs.getInt(key, 0);
     }
 }
